requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
tzdata==2022.1
asgiref==3.5.2
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
distlib==0.3.4
Django==4.0.4
django-ckeditor==6.4.1
django-js-asset==2.0.0
filelock==3.4.0
Pillow==9.1.1
pipenv==2021.11.23
platformdirs==2.4.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.10.0
virtualenv-clone==0.5.7
whitenoise==6.1.0
